Add driver for NetApp Workflow Automation tool
This commit adds a driver for NetApp's Workflow Automation (WFA)
software. WFA is configured in similar fashion to other storage drivers.
Fulfillment of Manila's requests depend on the execution of WFA workflow
scripts. These scripts depend on a number of values which can be defined
in Manila's extra specs.

Implements blueprint add-driver-for-netapp-workflow-automation

Change-Id: If658114c2026517abb9d600ffb4bd0c26a7e005c