[CI Fix] Hotfix for stable/liberty jobs
Tempest is branchless, but highly tempestuous in itself.
We have had many breakages in the manila stable branch jobs
because of changes made to the clients that we use for tempest.
This led us to pin a particular commit of tempest. However, when
the changes made to tempest are taken to stable branches of
devstack in any form [1], our stable branch jobs are no longer
protected by the 'stable' commit that we checkout. Without
bringing in the horde of changes to our stable branch, we could
cherry-pick the offending tempest change [2] on top of the
existing 'stable' commit.

[1] I63658b8d8dfa999e0feb79f8f2968f2b32e3ff57
[2] Iab2e6e04b6c5795a4d0c8214564106525b942308

Closes-Bug: #1603275

Change-Id: I4248c704256fc790082f3ac2515c6445ac7aa78a